rinbot/requirements.txt
discord.py>=2.3.0
yt-dlp
Pillow
psutil
aiosqlite
aiohttp
